+++ lib/src/balance/model/balance.dart
@@ -1,10 +1,10 @@
-/// Class encapsulating a balance object.
+/// Class encapsulating a [Balance] object.
 class Balance {
-  /// The payment method. Possible values are 'prepaid' and 'postpaid'.
+  /// The payment method. Possible values are `prepaid` and `postpaid`.
   final String payment;
 
-  /// The payment type. Possible values are: 'credits', 'euros', 'pounds'
-  /// and 'dollars'. For all other supported currencies, an ISO-4217 code
+  /// The payment type. Possible values are: `credits`, `euros`, `pounds`
+  /// and `dollars`. For all other supported currencies, an [ISO-4217](https://www.wikipedia.org/wiki/ISO_4217) code
   /// is returned.
   final String type;
 
@@ -16,13 +16,10 @@ class Balance {
   Balance({this.payment, this.type, this.amount});
 
   /// Construct a balance object from a json object.
-  factory Balance.fromJson(Map<String, dynamic> json) {
-    if (json == null) {
-      return null;
-    }
-    return Balance(
-        payment: json['payment'].toString(),
-        type: json['type'].toString(),
-        amount: double.parse(json['amount']));
-  }
+  factory Balance.fromJson(Map<String, dynamic> json) => json == null
+      ? null
+      : Balance(
+          payment: json['payment'].toString(),
+          type: json['type'].toString(),
+          amount: double.parse(json['amount'].toString()));
 }