+++ lib/src/callflows/model/callflows.dart
@@ -1,18 +1,15 @@
 import 'callflow.dart';
 
-/// Class encapsulating a list of callflows.
+/// Class encapsulating a list of [Callflows].
 class Callflows {
-  /// Listing of all [callflows].
+  /// Listing of all [Callflows].
   final List<Callflow> callflows;
 
   /// Constructor.
   Callflows({this.callflows});
 
-  /// Construct a step object from a [json] object.
-  factory Callflows.fromJson(Map<String, dynamic> json) {
-    if (json == null) {
-      return null;
-    }
-    return Callflows(callflows: Callflow.fromJsonList(json['data']));
-  }
+  /// Construct a [Callflows] object from a [json] object.
+  factory Callflows.fromJson(Map<String, dynamic> json) => json == null
+      ? null
+      : Callflows(callflows: Callflow.fromJsonList(json['data']));
 }