+++ lib/src/calls/api_calls_service.dart
@@ -1,13 +1,11 @@
 import 'dart:convert';
 
-import 'package:messagebird_dart/src/general/model/api_error.dart';
-
 import '../base_service.dart';
 import 'calls_service.dart';
 import 'model/call.dart';
 import 'model/calls.dart';
 
-/// API implementation of callflows service.
+/// API implementation of [CallsService].
 class ApiCallsService extends BaseService implements CallsService {
   /// Constructor.
   ApiCallsService(String accessKey, {int timeout, List<String> features})
@@ -17,46 +15,23 @@ class ApiCallsService extends BaseService implements CallsService {
   Future<Call> create(Map<String, dynamic> parameters) async {
     final response = await post('/calls',
         hostname: BaseService.voiceEndpoint, body: parameters);
-    if (response.statusCode == 201) {
-      return Future.value(Call.fromJson(json.decode(response.body)));
-    } else {
-      final ApiError error = ApiError.fromJson(json.decode(response.body));
-      throw Exception(error.code);
-    }
+    return Future.value(Call.fromJson(json.decode(response.body)));
   }
 
   @override
   Future<Calls> list() async {
     final response = await get('/calls', hostname: BaseService.voiceEndpoint);
-    if (response.statusCode == 200) {
-      return Future.value(Calls.fromJson(json.decode(response.body)));
-    } else {
-      final ApiError error = ApiError.fromJson(json.decode(response.body));
-      throw Exception(error.code);
-    }
+    return Future.value(Calls.fromJson(json.decode(response.body)));
   }
 
   @override
   Future<Call> read(int callId) async {
     final response =
         await get('/calls/$callId', hostname: BaseService.voiceEndpoint);
-    if (response.statusCode == 200) {
-      return Future.value(Call.fromJson(json.decode(response.body)));
-    } else {
-      final ApiError error = ApiError.fromJson(json.decode(response.body));
-      throw Exception(error.code);
-    }
+    return Future.value(Call.fromJson(json.decode(response.body)));
   }
 
   @override
-  Future<void> remove(int callId) async {
-    final response =
-        await delete('/calls/$callId', hostname: BaseService.voiceEndpoint);
-    if (response.statusCode == 200) {
-      return;
-    } else {
-      final ApiError error = ApiError.fromJson(json.decode(response.body));
-      throw Exception(error.code);
-    }
-  }
+  Future<void> remove(int callId) =>
+      delete('/calls/$callId', hostname: BaseService.voiceEndpoint);
 }