+++ lib/src/contacts/contacts_service.dart
@@ -1,29 +1,31 @@
-import 'package:http/http.dart' show Response;
+import 'model/contact.dart';
+import 'model/contacts.dart';
+import 'model/groups.dart';
+import 'model/messages.dart';
 
 /// Contacts service interface.
 abstract class ContactsService {
-  /// Create a new contact. [parameters] is optional.
-  Future<Response> create(String phoneNumber,
-      {Map<String, dynamic> parameters});
+  /// Create a new [Contact]. [parameters] is optional.
+  Future<Contact> create(String phoneNumber, {Map<String, dynamic> parameters});
 
-  /// Lists existing contacts. Pagination is optional. If a [limit] is set, an
+  /// Lists existing [Contacts]. Pagination is optional. If a [limit] is set, an
   /// [offset] is also required.
-  Future<Response> list({int limit, int offset});
+  Future<Contacts> list({int limit, int offset});
 
-  /// Lists the groups a contact is part of.
-  Future<Response> listGroups(String contactId, {int limit, int offset});
+  /// Lists the [Groups] a contact is part of.
+  Future<Groups> listGroups(String contactId, {int limit, int offset});
 
-  /// Lists the messages for a contact.
-  Future<Response> listMessages(String contactId, {int limit, int offset});
+  /// Lists the [Messages] for a contact.
+  Future<Messages> listMessages(String contactId, {int limit, int offset});
 
-  /// View an existing contact.
-  Future<Response> read(String id);
+  /// View an existing [Contact].
+  Future<Contact> read(String id);
 
-  /// Removes an existing contact. The Future completes with an error if
+  /// Removes an existing [Contact]. The Future completes with an error if
   /// applicable, but the data will never contain anything meaningful as the
   /// API returns an empty response for successful deletes.
-  Future<Response> remove(String id);
+  Future<void> remove(String id);
 
-  /// Updates an existing contact. [parameters] is optional.
-  Future<Response> update(int id, {Map<String, dynamic> parameters});
+  /// Updates an existing [Contact]. [parameters] is optional.
+  Future<Contact> update(int id, {Map<String, dynamic> parameters});
 }