+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+- Added model for contacts
+- Initial model for conversations
+
 ## 0.2.0
 - Completed all endpoint implementations
 - Initial model (balance, callflows & calls)